# MyBank Project/Documents/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/update/status":{"put":{"tags":["update-status-controller"],"operationId":"updateStatus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DebitCard"}}},"required":true},"responses":{"400":{"description":"Failed to block","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"404":{"description":"Customer not found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"DebitCard blocked successfully","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/bank/register":{"post":{"tags":["card-security-api"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardSecurity"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardSecurity"}}}}}}}},"components":{"schemas":{"DebitCard":{"required":["accountNumber","customerId","debitCardCvv","debitCardExpiry","debitCardNumber","debitCardPin","debitCardStatus","domesticLimit","internationalLimit"],"type":"object","properties":{"debitCardNumber":{"type":"integer","format":"int64"},"accountNumber":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int32"},"debitCardCvv":{"type":"integer","format":"int32"},"debitCardPin":{"type":"integer","format":"int32"},"debitCardExpiry":{"type":"string","format":"date-time"},"debitCardStatus":{"pattern":"^(block)$","type":"string"},"domesticLimit":{"type":"number","format":"double"},"internationalLimit":{"type":"number","format":"double"}}},"CardSecurity":{"type":"object","properties":{"customerName":{"type":"string"},"customerAddress":{"type":"string"},"customerStatus":{"type":"string"},"customerContact":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"maxAttempt":{"type":"integer","format":"int32"},"attempts":{"type":"integer","format":"int32"},"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}}}}}